requirements.txt
paho-mqtt==2.0.0
pydantic==2.7.1
CoolProp==6.4.3
polars==0.20.21
